refactor(interp): give the JIT integration its own file
interp.go held three separable things at ~3,000 lines: the threaded
interpreter's dispatch and heap machinery, the public Interpreter API, and
about a thousand lines of JIT integration.

That mixing had a visible symptom. compile_test.go existed with no compile.go
to own it, which is the concept file §12.1 forbids, and the previous change
parked it there as an exception. The production file should exist, because the
responsibility does.

interp.go is 2,071 lines now. jit.go holds tier-up event counting, compilation
requests, installation, the native dispatch wrappers, retirement,
deoptimization, and journal marshalling. tier.go keeps owning the policy those
consume. jit.go, jit_arm64.go, jit_stub.go, and jit_test.go read as one family.

Membership was decided per method by following callers, and three went the
non-obvious way. safepoint stays: it coordinates cancellation, fuel, the user
hook, and profiler sampling, and adopting cached modules is one of several
things it dispatches to, not its rule. restore stays: its other caller is
safepoint repairing frame fields the native wrappers null out, and its own body
touches no jit or journal type. entered, hit, and backedge move despite being
called from generated threaded dispatch, because their whole body is tier-up
counting - the generated code only holds a pointer to them.

globalKinds and function moved next to their only callers rather than being
appended in place; they had drifted hundreds of lines away.

Pure relocation. No renamed method, no changed signature.
