+++ docs/testing.md
@@ -45,7 +45,7 @@ condition that removes it (`docs/coding-patterns.md` §1.1).
 | File | Why it is white-box | Removal condition |
 |---|---|---|
 | `interp/tier_test.go` | Ties private `nativeFrameLimit` (declared in `interp/tier.go`) to the ARM64 invoke trampoline's stack reserve and total frame size via `arm64.StackReserve`/`arm64.FrameSize`, which own the byte arithmetic (`internal/asm/arm64/stack.go`), and reads `abi_arm64.s`'s two literals directly to compare against them. The complementary half of that reserve invariant — that `abi_arm64.s`'s own two literals agree with each other — needs no private interp state and lives as an ordinary external test, `arm64.TestFrameSize` (`internal/asm/arm64/stack_test.go`). | None known. `nativeFrameLimit` is interp-private bookkeeping with no contract to move with the ARM64 backend, which already lives in `internal/jit/arm64`. |
-| `interp/compile_test.go` | Drives `jit.Compiler.Compile` against a snapshot only the private `Interpreter.compileSnapshot` can build (module, constants, globals, heap, declared types, and the `HostStruct`/`field`/`conversion`/`coroutine` layout offsets `jit.Layout` needs), installs the result with the private `Interpreter.install`, and calls the compiled callable directly with `journalPtr`, asserting the native trap encoding (`journal.CellTrap`, `journal.CellExitID`, `jit.ExitDescriptor`) and splicing frame state to force a specific exit; also reads install bookkeeping (`tried`, `exits`) and `tracer.headers` that no metric separates from "not attempted" (`TestARM64_Backedge`). | None known. `newCompiler`, `compileSnapshot`, `journalPtr`, `install`, `tried`, `exits`, and the loop/call dispatch wrappers are interp-private mechanics with no contract to move with the ARM64 backend, which already lives in `internal/jit/arm64`. |
+| `interp/jit_test.go` | Drives `jit.Compiler.Compile` against a snapshot only the private `Interpreter.compileSnapshot` can build (module, constants, globals, heap, declared types, and the `HostStruct`/`field`/`conversion`/`coroutine` layout offsets `jit.Layout` needs), installs the result with the private `Interpreter.install`, and calls the compiled callable directly with `journalPtr`, asserting the native trap encoding (`journal.CellTrap`, `journal.CellExitID`, `jit.ExitDescriptor`) and splicing frame state to force a specific exit; also reads install bookkeeping (`tried`, `exits`) and `tracer.headers` that no metric separates from "not attempted" (`TestARM64_Backedge`). | None known. `newCompiler`, `compileSnapshot`, `journalPtr`, `install`, `tried`, `exits`, and the loop/call dispatch wrappers are interp-private mechanics with no contract to move with the ARM64 backend, which already lives in `internal/jit/arm64`. |
 | `interp/trace_test.go` | Calls `tracer.capture` without `Run`, which is the only way to prove speculative capture snapshots a container instead of mutating the live heap. Any public path runs the real instructions too, so the isolation claim is unobservable by construction. | None known. The recorder stays in `interp`, and the claim is only provable from inside it. |
 
 ### Known Coverage Gaps
@@ -56,7 +56,7 @@ proxy double (`docs/coding-patterns.md` §12.2, §12.3).
 
 | Uncovered | Why it cannot be reached publicly |
 |---|---|
-| `Interpreter.retire` and the watchdog (`retireWindow`, `retireGiveUpThreshold` in `interp/tier.go`; `checkRetire` in `interp/interp.go`) | Retirement is not observable from outside. A program producing a `trace-cut` give-up on every native entry runs past 80,000 entries without native-entry counts plateauing, so no public metric distinguishes a retired entry from a live one. The mechanism is what handled the RecursiveFib/35 regression, where a native entry was a net loss; a break in the give-up accounting would not fail any test today. |
+| `Interpreter.retire` and the watchdog (`retireWindow`, `retireGiveUpThreshold` in `interp/tier.go`; `checkRetire` in `interp/jit.go`) | Retirement is not observable from outside. A program producing a `trace-cut` give-up on every native entry runs past 80,000 entries without native-entry counts plateauing, so no public metric distinguishes a retired entry from a live one. The mechanism is what handled the RecursiveFib/35 regression, where a native entry was a net loss; a break in the give-up accounting would not fail any test today. |
 | Hot-entry counter saturation | The counter and the tier-up trigger can only reach their overflow edge by being written directly. |
 | Trace-tree attribution to the true entry IP | Requires driving compilation at a fabricated frame IP. |
 | Dataflow fact widening at a control-flow join (`mergeSlot` in `internal/jit`) | A slot's `refKnown`/`calleeKnown` facts are planning-internal: the backend never reads them, so nothing exports them. Their only external effect is which plans a frontend produces or rejects, which no fixture isolates from the rest of planning. |
